focussing on height adjustment for nested tables in columns
Need to refactor cardinality calc, as that's still messed up from other
refactoring to layout decoupling

+++ kramer/src/main/java/com/chiralbehaviors/layout/ColumnSet.java
@@ -66,7 +66,7 @@ public class ColumnSet {
         }
 
         // compression
-        double columnWidth = relax(justified / count);
+        double columnWidth = relax(justified / (double) count);
         firstColumn.setWidth(columnWidth);
         double compressed = relax(columnWidth - labelWidth);
         firstColumn.getFields()
