+++ kramer/src/main/java/com/chiralbehaviors/layout/RelationLayout.java
@@ -86,7 +86,7 @@ public class RelationLayout extends SchemaNodeLayout {
     public void adjustHeight(double delta) {
         super.adjustHeight(delta);
         if (useTable) {
-            double subDelta = delta / children.size();
+            double subDelta = delta / (double) children.size();
             if (delta >= 1.0) {
                 children.forEach(f -> f.adjustHeight(subDelta));
             }
